requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.2
python-dotenv==1.0.1
websockets>=13.0

# Telegram Bot
python-telegram-bot==21.6

# Document Processing
pdfplumber==0.11.4
PyMuPDF==1.24.10
python-docx==1.1.2
Pillow==10.4.0
openpyxl==3.1.5

# AI - Google Gemini (free tier)
google-generativeai==0.8.3

# TTS - FPT.AI
aiofiles==24.1.0

# Database
supabase==2.9.1

# Storage
boto3==1.35.0

# Utils
tenacity==9.0.0
